chore(model-server): revert previously introduced openapi gradle subproject
