+++ gradle/libs.versions.toml
@@ -81,7 +81,6 @@ ktor-serialization-json = { group = "io.ktor", name = "ktor-serialization-kotlin
 keycloak-authz-client = { group = "org.keycloak", name = "keycloak-authz-client", version = "23.0.3" }
 guava = { group = "com.google.guava", name = "guava", version = "32.1.3-jre" }
 org-json = { group = "org.json", name = "json", version = "20231013" }
-google-gson = { group = "com.google.code.gson", name = "gson", version = "2.10.1" }
 google-oauth-client = { group = "com.google.oauth-client", name = "google-oauth-client", version = "1.34.1" }
 google-oauth-clientjetty = { group = "com.google.oauth-client", name = "google-oauth-client-jetty", version = "1.34.1" }
