+++ model-server-openapi/specifications/model-server-deprecated.yaml
