+++ model-server-openapi/specifications/model-server-light.yaml
