+++ model-server-openapi/specifications/model-server-operative.yaml
