+++ model-server-openapi/specifications/model-server.yaml
