+++ model-server/src/main/resources/openapi/templates/AppMain.kt.mustache
