+++ model-server/src/main/resources/openapi/templates/Paths.kt.mustache
