+++ model-server/src/main/resources/openapi/templates/api.mustache
