+++ model-server/src/main/resources/openapi/templates/data_class.mustache
